requirements.txt
streamlit-nightly
pandas
gtts